Bump CSV and Image to 0.1.0
Bump the CSV to 0.1.0 and also use version 0.1.0 of the container
image to match the certified ones.

Add more details to the CSV description, and some metadata
specific to Openshift Marketplace.

Finally, this patch also add a convenient log statement for the
actual namespace that will be monitored by the configmap controller.