Add github workflows for release
This patch adds workflows for automatically pushing new image tags
when a release is made, and uploading operator yaml files as release
assets.

In addition, it modifies the build workflow to automatically update
the 'latest' tag for the operator image.

In order to support these workflows the patch also includes:
- Scripts for generating assets to upload
- addition of a VERSION file to mark the current version
- Minor changes to the Makefile, mostly for tracking docker image versions