Immediately respond when the customer correct error in role
The previous error handling logic is that when the operator hit
error during applying configuration, the reconciler will receive
an error then requeue the request. However, the requeue mechanism
has a potential backoff, that caused the operator may not
immediately respond when the customer correct error in operator's
cluster role.

This patch will additional watch the cluster role change for
the immediate response.