Improvements for SecretNotEqual function
Improve logic for comparing keys in the secret in order to avoid
conditions where empty and missing tls keys (eg.: tls.ca) are interpreted
as differences in the secret.

Also, do not use DeepEqual as there is only need for comparing tls data
in the secret. This will avoid unnecessarily restarting ncp for unrelated
operator secret changes.

Finally do not include tls.ca in the projected volume if not specified in
the operator secret, or empty.

This commit also adds unit tests for secret comparison cases not
previously covered