Mount nmcli to avoid the version conflicts between host and OVS container (#289)
We are planning to upgrade the NCP base images from UBI 8 to UBI 9.
In UBI 9, the NetworkManager version is newer than the host, that will
cause the issue that the nmcli cannot work in the OVS container.
So this patch will mount the host's nmcli to overwrite the existing one
in the OVS container. In the future, we can try to not install
NetworkManager in the container.

Testing done:
CIBot passed for UBI 9. (testrunid=4077352)